[UI] Phase 4: move registry UI to ui/components/registry (#18747)
Lift the registry-related UI out of Settings so the registry domain is
no longer buried under the legacy Settings umbrella. This is a pure
structural relocation: file contents are unchanged except for import
paths that needed to adapt to the new depth, and the existing top-level
`Registry/RegistryModal.tsx` is consolidated into the same lowercase
`registry/` directory so the registry surface has a single home.

- git mv ui/components/Settings/Registry/* -> ui/components/registry/*
- consolidate ui/components/Registry/RegistryModal.tsx -> ui/components/registry/RegistryModal.tsx
- update importers in Header.tsx, NavigatorExtension.tsx,
  Dashboard/view-component.tsx, and Settings/MesherySettings.tsx
- update relative paths inside moved files to reflect the new depth
- update eslint.config.js legacy offender allowlists to track the new
  paths

Phase 2 #18735 already converted these files to Sistent/@sistent/sistent,
so this PR carries no MUI churn.

Fixes #18747
Refs #18659, #18654
