+++ ui/components/registry/MeshModel.style.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { styled, Button, alpha } from '@sistent/sistent';
-import { TreeItem, treeItemClasses } from '../../shared/TreeView';
+import { TreeItem, treeItemClasses } from '../shared/TreeView';
 
 export const DisableButton = styled(Button)(({ theme }) => ({
   '&.MuiButtonBase-root:disabled': {
