+++ ui/components/registry/RelationshipTree.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { SimpleTreeView } from '../../shared/TreeView';
+import { SimpleTreeView } from '../shared/TreeView';
 import { CircularProgress } from '@sistent/sistent';
 import { RELATIONSHIPS } from '@/constants/navigator';
-import MinusSquare from '../../../assets/icons/MinusSquare';
-import PlusSquare from '../../../assets/icons/PlusSquare';
-import DotSquare from '../../../assets/icons/DotSquare';
+import MinusSquare from '../../assets/icons/MinusSquare';
+import PlusSquare from '../../assets/icons/PlusSquare';
+import DotSquare from '../../assets/icons/DotSquare';
 import StyledTreeItem from './StyledTreeItem';
 
 type RelationshipTreeProps = {
