+++ ui/components/registry/Stepper/style.ts
