+++ ui/components/registry/StyledTreeItem.tsx
