+++ ui/eslint.config.js
@@ -58,7 +58,7 @@ const legacyRestrictedImportOffenders = [
   'components/NotificationCenter/notificationCenter.style.tsx',
   'components/ReactSelectWrapper.tsx',
   'components/RelationshipBuilder/RelationshipFormStepper.tsx',
-  'components/Settings/Registry/Stepper/CSVStepper.tsx',
+  'components/registry/Stepper/CSVStepper.tsx',
   'components/workspaces/SpacesSwitcher/MainDesignsContent.tsx',
   'components/workspaces/SpacesSwitcher/WorkspaceModal.tsx',
   'components/workspaces/SpacesSwitcher/components.tsx',
@@ -121,8 +121,8 @@ const legacyLiteralColorOffenders = [
   'components/Performance/index.tsx',
   'components/Performance/style.tsx',
   'components/Settings/MesherySettings.tsx',
-  'components/Settings/Registry/MeshModelDetails.tsx',
-  'components/Settings/Registry/helper.ts',
+  'components/registry/MeshModelDetails.tsx',
+  'components/registry/helper.ts',
   'components/workspaces/SpacesSwitcher/WorkspaceSwitcher.tsx',
   'components/workspaces/SpacesSwitcher/styles.tsx',
   'components/TypingFilter/style.tsx',
@@ -165,7 +165,7 @@ const legacyMaxLineSoftOffenders = [
   'components/Lifecycle/Environments/index.tsx',
   'components/Navigator.tsx',
   'components/Performance/PerformanceResults.tsx',
-  'components/Settings/Registry/Stepper/UrlStepper.tsx',
+  'components/registry/Stepper/UrlStepper.tsx',
   'components/UserPreferences/index.tsx',
   'components/connections/meshSync/index.tsx',
   'components/telemetry/grafana/GrafanaCustomChart.tsx',
@@ -284,25 +284,25 @@ const legacyInlineStyleOffenders = [
   'components/Performance/assets/twitterIcon.tsx',
   'components/Performance/index.tsx',
   'components/ReactSelectWrapper.tsx',
-  'components/Registry/RegistryModal.tsx',
   'components/RelationshipBuilder/CreateRelationshipModal.tsx',
   'components/RelationshipBuilder/RelationshipFormStepper.tsx',
   'components/Settings/MesherySettings.tsx',
   'components/Settings/MesherySettingsPerformanceComponent.tsx',
-  'components/Settings/Registry/ComponentTree.tsx',
-  'components/Settings/Registry/CreateModelModal.tsx',
-  'components/Settings/Registry/ImportModel.tsx',
-  'components/Settings/Registry/ImportModelModal.tsx',
-  'components/Settings/Registry/MeshModelComponent.tsx',
-  'components/Settings/Registry/MeshModelDetails.tsx',
-  'components/Settings/Registry/MesheryTreeView.tsx',
-  'components/Settings/Registry/MesheryTreeViewItem.tsx',
-  'components/Settings/Registry/MesheryTreeViewModel.tsx',
-  'components/Settings/Registry/MesheryTreeViewRegistrants.tsx',
-  'components/Settings/Registry/RelationshipTree.tsx',
-  'components/Settings/Registry/Stepper/CSVStepper.tsx',
-  'components/Settings/Registry/Stepper/UrlStepper.tsx',
-  'components/Settings/Registry/StyledTreeItem.tsx',
+  'components/registry/ComponentTree.tsx',
+  'components/registry/CreateModelModal.tsx',
+  'components/registry/ImportModel.tsx',
+  'components/registry/ImportModelModal.tsx',
+  'components/registry/MeshModelComponent.tsx',
+  'components/registry/MeshModelDetails.tsx',
+  'components/registry/MesheryTreeView.tsx',
+  'components/registry/MesheryTreeViewItem.tsx',
+  'components/registry/MesheryTreeViewModel.tsx',
+  'components/registry/MesheryTreeViewRegistrants.tsx',
+  'components/registry/RegistryModal.tsx',
+  'components/registry/RelationshipTree.tsx',
+  'components/registry/Stepper/CSVStepper.tsx',
+  'components/registry/Stepper/UrlStepper.tsx',
+  'components/registry/StyledTreeItem.tsx',
   'components/workspaces/SpacesSwitcher/DesignViewListItem.tsx',
   'components/workspaces/SpacesSwitcher/MainDesignsContent.tsx',
   'components/workspaces/SpacesSwitcher/MainViewsContent.tsx',
